refactor: split stream.py god module into focused sub-modules (#247)
