+++ silas/core/stream/__init__.py
@@ -1,0 +1,6 @@
+"""Stream package — re-exports for backward compatibility."""
+
+from silas.core.stream._nonce import _InMemoryNonceStore
+from silas.core.stream._stream import Stream, TurnProcessor
+
+__all__ = ["Stream", "TurnProcessor", "_InMemoryNonceStore"]
